refactor(sdk): extract auth + DPoP + headers to _client/_auth.py
PR 8/N of the SDK split. Moves the security-CRITICAL Auth + DPoP +
headers domain out of the ~2 kLOC ``cullis_sdk/client.py`` into
``cullis_sdk/_client/_auth.py``:

* ``_dpop_proof``, ``_update_nonce``, ``_headers``, ``_authed_request``
  (RFC 9449 proof + nonce handling, broker-direct deprecation warning,
  PR #690 auto re-login on token-expiry 401, recursion guard).
* ``login``, ``login_from_pem``, ``_legacy_login_from_pem``
  (legacy BYOCA → Court /v1/auth/token, deprecation warning fires once).
* ``login_via_proxy``, ``login_via_proxy_with_local_key``
  (ADR-004 §2.4 + Tech-debt #2 proxy-mediated login flows).

Movement only — byte-perfect. Order of POSTs to /v1/auth/token, the
``use_dpop_nonce`` 401 replay branch, the ADR-009 Phase 2
``X-Cullis-Mastio-Signature`` forwarding, and the
``self._relogin_callable`` wiring on success are preserved verbatim.

Module-level helpers (``_check_insecure_tls``,
``_build_proxy_http_client``, ``_cert_key_pair_matches``,
``InsecureTLSWarning``) intentionally stay in ``cullis_sdk.client``
because tests (``test_sdk_tls_hardening``, ``test_sdk_proxy_client_ca_pin``,
``test_h10_ca_chain_threading``) import + ``mock.patch`` them at that
path, and ``cullis_sdk/_client/_rfq.py`` reuses the same lazy import.
Backward compat preserved.

``class CullisClient(...)`` now inherits from 8 mixins in alphabetical
order (``_AiGatewayMixin``, ``_AuthMixin``, ``_DiscoveryMixin``,
``_GuardianMixin``, ``_MessagingOneshotMixin``, ``_RfqMixin``,
``_SessionsMixin``, ``_WebSocketMixin``). Zero breaking change.

NO ``@deprecated`` decorator here — that lands in a dedicated PR once
the split is complete.

Verification
* ``pytest tests/ -n auto --dist=loadfile`` — 3494 passed, 38 skipped.
  Failures are pre-existing flakes documented in memory:
  2 postgres bindings + 3 audit race-safety (sqlite-concurrency).
* ``./sandbox/smoke.sh full`` — PASS=25 FAIL=0 SKIP=1.
  B3.5 / B3.7 / B3.8 (OIDC dashboard login) and B4.4 / B4.5 / B4.6
  (DPoP-bound API-key auth) exercise every login path on the moved
  surface — all green.

Follows the pattern of #702 (WebSocket), #706 (Guardian), #711
(Discovery), #713 (RFQ + notifications + network), #716 (AI gateway
+ MCP), #718 (Sessions), #720 (Messaging oneshot).
